outsource enums to separate files in every module
