+++ backend/dataall/modules/datasets/api/table/enums.py
@@ -1,0 +1,15 @@
+from dataall.base.api.constants import GraphQLEnumMapper
+
+
+class DatasetSortField(GraphQLEnumMapper):
+    created = 'created'
+    updated = 'updated'
+    label = 'label'
+
+
+class DatasetTablePreviewStatus(GraphQLEnumMapper):
+    QUEUED = 'QUEUED'
+    RUNNING = 'RUNNING'
+    SUCCEEDED = 'SUCCEEDED'
+    FAILED = 'FAILED'
+    CANCELLED = 'CANCELLED'
