+++ backend/dataall/modules/worksheets/api/resolvers.py
@@ -1,15 +1,9 @@
-from dataall.base.api.constants import GraphQLEnumMapper
 from dataall.base.db import exceptions
+from dataall.modules.worksheets.api.enums import WorksheetRole
 from dataall.modules.worksheets.db.models import Worksheet
 from dataall.modules.worksheets.db.repositories import WorksheetRepository
 from dataall.modules.worksheets.services.worksheet_services import WorksheetService
 from dataall.base.api.context import Context
-
-
-class WorksheetRole(GraphQLEnumMapper):
-    Creator = '950'
-    Admin = '900'
-    NoPermission = '000'
 
 
 def create_worksheet(context: Context, source, input: dict = None):
